reuse tests between hooks and hocs
